Extract combo pages display to helper
