+++ helpers/index.js
@@ -1,3 +1,5 @@
+const { MessageEmbed, MessageActionRow, MessageButton } = require('discord.js');
+
 //////////////////////////////////////////////////////
 // Cacheing 
 //////////////////////////////////////////////////////
@@ -5,6 +7,7 @@
 const fs = require('fs')
 const resolve = require('path').resolve;
 const gHelper = require('../helpers/general')
+const discordHelper = require('../helpers/discord.js');
 
 DIR1 = 'cache'
 DIR2 = 'index'
@@ -136,6 +139,126 @@ function altMapsFields(ogMapAbbr, allCompletedMapAbbrs, isWaterEntity) {
     }
 }
 
+const multipageButtons = new MessageActionRow().addComponents(
+    new MessageButton().setCustomId('-1').setLabel('⬅️').setStyle('PRIMARY'),
+    new MessageButton().setCustomId('1').setLabel('➡️').setStyle('PRIMARY')
+);
+
+async function displayOneOrMultiplePages(interaction, colData, setCustomFields) {
+    MAX_NUM_ROWS = 15;
+    const numRows = Object.values(colData)[0].length;
+    let leftIndex = 0;
+    let rightIndex = Math.min(MAX_NUM_ROWS - 1, numRows - 1);
+
+    /**
+     * creates embed for next page
+     * @param {int} direction
+     * @returns {MessageEmbed}
+     */
+    async function createPage(direction = 1) {
+        // The number of rows to be displayed is variable depending on the characters in each link
+        // Try 15 and decrement every time it doesn't work.
+        for (
+            maxNumRowsDisplayed = MAX_NUM_ROWS;
+            maxNumRowsDisplayed > 0;
+            maxNumRowsDisplayed--
+        ) {
+            let challengeEmbed = new Discord.MessageEmbed()
+
+            challengeEmbed.addField(
+                '# Combos',
+                `**${leftIndex + 1}**-**${rightIndex + 1}** of ${numRows}`
+            );
+
+            for (header in colData) {
+                const data =
+                    numRows <= maxNumRowsDisplayed
+                        ? colData[header]
+                        : colData[header].slice(leftIndex, rightIndex + 1);
+
+                challengeEmbed.addField(
+                    gHelper.toTitleCase(header.split('_').join(' ')),
+                    data.join('\n'),
+                    true
+                );
+            }
+
+            setCustomFields(challengeEmbed)
+
+            if (discordHelper.isValidFormBody(challengeEmbed)) {
+                return [challengeEmbed, numRows > maxNumRowsDisplayed];
+            }
+
+            if (direction > 0) rightIndex--;
+            if (direction < 0) leftIndex++;
+        }
+    }
+
+    async function displayPages(direction = 1) {
+        let [embed, multipage] = await createPage(direction);
+
+        await interaction.editReply({
+            embeds: [embed],
+            components: multipage ? [multipageButtons] : [],
+        });
+
+        if (!multipage) return;
+
+        const filter = (selection) => {
+            // Ensure user clicking button is same as the user that started the interaction
+            if (selection.user.id !== interaction.user.id) {
+                return false;
+            }
+            // Ensure that the button press corresponds with this interaction and wasn't
+            // a button press on the previous interaction
+            if (selection.message.interaction.id !== interaction.id) {
+                return false;
+            }
+            return true;
+        };
+
+        const collector = interaction.channel.createMessageComponentCollector({
+            filter,
+            componentType: 'BUTTON',
+            time: 20000
+        });
+
+        collector.on('collect', async (buttonInteraction) => {
+            collector.stop();
+            buttonInteraction.deferUpdate();
+
+            switch (parseInt(buttonInteraction.customId)) {
+                case -1:
+                    rightIndex = (leftIndex - 1 + numRows) % numRows;
+                    leftIndex = rightIndex - (MAX_NUM_ROWS - 1);
+                    if (leftIndex < 0) leftIndex = 0;
+                    await displayPages(-1);
+                    break;
+                case 1:
+                    leftIndex = (rightIndex + 1) % numRows;
+                    rightIndex = leftIndex + (MAX_NUM_ROWS - 1);
+                    if (rightIndex >= numRows) rightIndex = numRows - 1;
+                    await displayPages(1);
+                    break;
+            }
+        });
+
+        collector.on('end', async (collected) => {
+            if (collected.size == 0) {
+                await interaction.editReply({
+                    embeds: [embed],
+                    components: []
+                });
+            }
+        });
+    }
+
+    // Gets the reaction to the pagination message by the command author
+    // and respond by turning the page in the correction direction
+
+    await displayPages(1);
+}
+
 module.exports = {
     hasCachedCombos,
     fetchCachedCombos,
@@ -145,4 +268,5 @@ module.exports = {
     parseMapNotes,
 
     altMapsFields,
+    displayOneOrMultiplePages,
 }
